RenderLayer overhaul.
Added entry point for external code to render to shadow maps.
