+++ EvoEngine_SDK/src/CpuRayTracerCamera.cpp
@@ -36,12 +36,12 @@ void CpuRayTracerCamera::OnCreate() {
 
 void CpuRayTracerCamera::Capture(const CaptureParameters& parameters,
                                  const std::shared_ptr<Texture2D>& target_texture) const {
-  std::shared_ptr<RenderInstances> render_instances;
+  std::shared_ptr<RenderInstanceStorage> render_instances;
 
   if (const auto render_layer = Application::GetLayer<RenderLayer>()) {
-    render_instances = render_layer->render_instances_list[Platform::GetCurrentFrameIndex()];
+    render_instances = render_layer->GetCurrentRenderInstances();
   } else {
-    render_instances = std::make_shared<RenderInstances>();
+    render_instances = std::make_shared<RenderInstanceStorage>();
   }
   std::vector<glm::vec4> pixels{resolution.x * resolution.y};
   CameraInfoBlock camera_info_block;
